pyproject.toml
[project]
name = "spotify-to-beatport"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "nexus106", email = "106nexus@gmail.com" }
]
dependencies = [
    "spotipy>=2.24.0",
    "python-dotenv>=1.0.1",
    "bs4>=0.0.2",
    "requests>=2.32.3",
    "lxml>=5.2.2",
    "selenium>=4.22.0",
    "chromedriver-binary==126.0.6478.126.0",
    "mypy>=1.10.1",
    "types-selenium>=3.141.9",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/spotify_to_beatport"]
